+++ resources/js/Pages/Schedule/Index.vue
@@ -29,78 +29,6 @@
 
     <ScheduleGridContainer />
 
-<!--    <div class="ml-5 px-2 mb-2">-->
-<!--      <span class="text-sm uppercase text-purple-500">All times are listed in your timezone.</span>-->
-<!--    </div>-->
-
-<!--    <div class="schedule-grid">-->
-<!--    <div class="header-row">-->
-<!--      &lt;!&ndash; Time slots header &ndash;&gt;-->
-<!--      <div class="time-cell bg-gray-900 px-3 py-2 text-center font-bold" v-for="interval in nextFourHoursWithHalfHourIntervals" :key="interval.dateTime">-->
-<!--        {{ interval.formatted }}-->
-<!--      </div>-->
-<!--    </div>-->
-<!--    </div>-->
-<!--    <ScheduleGrid />-->
-
-
-<!--    <div class="schedule-grid">-->
-<!--      <div class="header-row">-->
-<!--        &lt;!&ndash; Time slots header &ndash;&gt;-->
-<!--        <div class="time-cell bg-gray-900 px-3 py-2 text-center font-bold" v-for="interval in nextFourHoursWithHalfHourIntervals" :key="interval.dateTime">-->
-<!--          {{ interval.formatted }}-->
-<!--        </div>-->
-<!--      </div>-->
-
-<!--      <div v-if="scheduleStore.isLoading && scheduleStore.nextFourHoursOfContent.length === 0"-->
-<!--            class="w-full flex justify-center text-center items-center">-->
-<!--        <span class="loading loading-ball loading-xl text-info"></span>-->
-<!--      </div>-->
-
-<!--      <div class="content-row">-->
-<!--        &lt;!&ndash; Scheduled shows &ndash;&gt;-->
-<!--        <template v-for="item in nextFourHoursOfContent" :key="item.id">-->
-<!--          <div-->
-<!--              class="show-cell flex flex-col w-full h-full"-->
-<!--              :class="getCellClasses(item.type)"-->
-<!--              :style="gridPlacement(item.start_time, item.durationMinutes)"-->
-<!--              @click="handleShowClick(item)"-->
-<!--          >-->
-<!--            <div  class="item-content flex flex-col items-center justify-center gap-y-2 flex-grow">-->
-<!--              <h3>{{ item.content.show?.name || 'No Show Name' }}</h3>-->
-<!--              &lt;!&ndash; Display the image if available &ndash;&gt;-->
-<!--              <SingleImage v-if="item.content.show?.image"-->
-<!--                           :image="item.content.show?.image"-->
-<!--                           :alt="item.content.show?.name"-->
-<!--                           class="content-image"/>-->
-<!--              &lt;!&ndash; Fallback placeholder if no image &ndash;&gt;-->
-<!--              <div v-else class="placeholder"></div>-->
-<!--            </div>-->
-<!--          </div>-->
-<!--        </template>-->
-<!--      </div>-->
-
-
-<!--      &lt;!&ndash; Status Row &ndash;&gt;-->
-<!--      <div class="status-row">-->
-<!--        &lt;!&ndash; Status for each show &ndash;&gt;-->
-<!--        <template v-for="(item, index) in nextFourHoursOfContent" :key="`status-${item.id}`">-->
-<!--          <div-->
-<!--              :class="getStatusCellClasses(index)" :style="gridPlacement(item.start_time, item.durationMinutes)"-->
-<!--          >-->
-<!--            &lt;!&ndash; Use the index to determine the status &ndash;&gt;-->
-<!--            <span v-if="index === 0">NOW PLAYING</span>-->
-<!--            <span v-else-if="index === 1">COMING UP NEXT</span>-->
-
-<!--          </div>-->
-<!--        </template>-->
-<!--      </div>-->
-
-
-<!--    </div>-->
-
-
-
     <PopUpModal :id="`goToNowPlayingModal`">
       <template v-slot:header>Now Playing</template>
       <template v-slot:main><span class="text-orange-500">This modal is temporary. This will take you to the now playing show or episode page.</span>
@@ -112,13 +40,6 @@
       </template>
     </PopUpModal>
 
-<!--    <div class="bg-gray-600 rounded-lg shadow m-10 p-4">-->
-
-
-<!--      <TodayView/>-->
-<!--    </div>-->
-
-
   </div>
 </template>
 
@@ -129,8 +50,6 @@ import { useUserStore } from '@/Stores/UserStore'
 import { useScheduleStore } from '@/Stores/ScheduleStore'
 import Message from '@/Components/Global/Modals/Messages'
 import PopUpModal from '@/Components/Global/Modals/PopUpModal'
-import TodayView from '@/Components/Global/Calendar/TodayView.vue'
-// import ScheduleGrid from '@/Components/Pages/Schedule/ScheduleGrid.vue'
 import ScheduleGridContainer from '@/Components/Global/Schedule/ScheduleGridContainer.vue'
 import { onMounted, ref, watch } from 'vue'
 import dayjs from 'dayjs';
@@ -145,25 +64,25 @@ let props = defineProps({
   can: Object,
 })
 
-onMounted(() => {
-  scheduleStore.initializeTimeSlots();
-});
+// onMounted(() => {
+//   scheduleStore.initializeTimeSlots();
+// });
 
 // Define a reactive watcher on the timezone
 // This watcher will call fetchSchedules whenever the timezone changes and is not null
-watch(
-    () => userStore.timezone,
-    async (newTimezone, oldTimezone) => {
-      // Ensure the timezone is set before calling preloadWeeklyContent
-      if (newTimezone) {
-        const startDate = dayjs();
-        const endDate = startDate.add(4, 'hour');
-        await scheduleStore.fetchSchedules(startDate.format(), endDate.format());
-        console.log('Preload schedules...')
-      }
-    },
-    {immediate: true}, // This option ensures the watcher is triggered immediately on mount
-)
+// watch(
+//     () => userStore.timezone,
+//     async (newTimezone, oldTimezone) => {
+//       // Ensure the timezone is set before calling preloadWeeklyContent
+//       if (newTimezone) {
+//         const startDate = dayjs();
+//         const endDate = startDate.add(4, 'hour');
+//         await scheduleStore.fetchSchedules(startDate.format(), endDate.format());
+//         console.log('Preload schedules...')
+//       }
+//     },
+//     {immediate: true}, // This option ensures the watcher is triggered immediately on mount
+// )
 
 function openModal(modalName) {
   document.getElementById(modalName).showModal()